Check before detaching categories if deleted entity was soft deleted (#29)

+++ src/Traits/Categorizable.php
@@ -89,7 +89,8 @@ trait Categorizable
     public static function bootCategorizable()
     {
         static::deleted(function (self $model) {
-            $model->categories()->detach();
+            // Check if this is a soft delete or not by checking if `SoftDeletes::isForceDeleting` method exists
+            (method_exists($model, 'isForceDeleting') && ! $model->isForceDeleting()) || $model->categories()->detach();
         });
     }
